+++ examples/full/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 
 const { ServiceBroker } = require("moleculer");
-const redisAdapter = require("socket.io-redis");
+// const { createAdapter } = require("@socket.io/redis-adapter");
 const SocketIOService = require("../../");
 const express = require("express");
 const fs = require("fs");
@@ -94,7 +94,8 @@ broker.createService({
 		save: {
 			handler(ctx) {
 				return new this.Promise((resolve, reject) => {
-					const filePath = path.join(__dirname, "public/upload", ctx.meta.filename);
+					const safeFilename = path.basename(ctx.params.filename);
+					const filePath = path.join(__dirname, "public/upload", safeFilename);
 					const f = fs.createWriteStream(filePath);
 					f.on("close", () => {
 						this.logger.info(`Uploaded file stored in '${filePath}'`);
@@ -102,7 +103,7 @@ broker.createService({
 					});
 					f.on("error", err => reject(err));
 
-					ctx.params.pipe(f);
+					ctx.stream.pipe(f);
 				});
 			}
 		}
@@ -148,11 +149,11 @@ const ioService = broker.createService({
 							let stream = new Duplex();
 							stream.push(file);
 							stream.push(null);
-							await this.$service.broker.call("file.save", stream, {
-								meta: {
-									filename: name
-								}
-							});
+							await this.$service.broker.call(
+								"file.save",
+								{ filename: name },
+								{ stream }
+							);
 							respond(null, name);
 						}
 					}